Refactor tests; prepare for wrapper class
