+++ src/test/java/io/spokestack/spokestack/tts/SpokestackTTSOutputTest.java
@@ -11,8 +11,13 @@ import androidx.lifecycle.LifecycleOwner;
 import androidx.lifecycle.LifecycleRegistry;
 import com.google.android.exoplayer2.ExoPlaybackException;
 import com.google.android.exoplayer2.ExoPlayer;
+import com.google.android.exoplayer2.PlaybackParameters;
 import com.google.android.exoplayer2.Player;
+import com.google.android.exoplayer2.Timeline;
 import com.google.android.exoplayer2.source.MediaSource;
+import com.google.android.exoplayer2.source.SinglePeriodTimeline;
+import com.google.android.exoplayer2.source.TrackGroupArray;
+import com.google.android.exoplayer2.trackselection.TrackSelectionArray;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -192,16 +197,20 @@ public class SpokestackTTSOutputTest {
 
         // these methods are implemented solely to maintain compatibility with
         // older Android APIs; calling them should do nothing
-        ttsOutput.onTimelineChanged(null, 0);
-        ttsOutput.onTimelineChanged(null, null, 0);
-        ttsOutput.onTracksChanged(null, null);
+        Timeline timeline = new SinglePeriodTimeline(0, false, false, false);
+        ttsOutput.onTimelineChanged(timeline, 0);
+        ttsOutput.onTimelineChanged(timeline, null, 0);
+        TrackGroupArray groupArray = new TrackGroupArray();
+        TrackSelectionArray selectionArray = new TrackSelectionArray();
+        ttsOutput.onTracksChanged(groupArray, selectionArray);
         ttsOutput.onLoadingChanged(true);
         ttsOutput.onPlaybackSuppressionReasonChanged(0);
         ttsOutput.onIsPlayingChanged(false);
         ttsOutput.onRepeatModeChanged(0);
         ttsOutput.onShuffleModeEnabledChanged(false);
         ttsOutput.onPositionDiscontinuity(-10);
-        ttsOutput.onPlaybackParametersChanged(null);
+        PlaybackParameters params = new PlaybackParameters(1.0f);
+        ttsOutput.onPlaybackParametersChanged(params);
         ttsOutput.onSeekProcessed();
     }
